scale: Provide median-based filtering
Users have reported that the previous app's SAW behavior
would trigger SAW prematurely when working with a pitcher
rinser or knockbox at the same time.

In order to try to mitigate the effects of vibration,
without introducing too much delay or bias, a median filter
is provided for mass-flow estimation in addition to the one
already being used for post-flow weight estimation.

A median filter selects the "middle" value from the series,
helpful for "ignoring" deviations, such as vibrations.

One drawback is that there is slightly more delay.
With the current parameters, add about 0.5 seconds
of delay to the weight line, and the flow lags by
about 0.2 seconds over the about 0.5 second delay
of the "finite difference" (default) algorithm
recently introduced.

These compensation factors are already included
in the SAW logic, though a spot check is suggested.

Users can enable/disable the use of these filters with

$::settings(high_vibration_scale_filtering)

which accepts True (or False) (or other Tcl Booleans)

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>