scale: Reset estimator data on scale tare seen
When a tare request is sent from the app it is often ~350 ms
until the received weights are "0" and can be over 500 ms.
Previously, these "old" weights were used in the mass and
mass-flow estimates. With estimation windows around 1 second,
these could creep into the "active" portion of a shot record.

Now ::device::scale:on_tare_seen is called if a "0" weight update
is received from the scale witin _tare_awaiting_zero_ms (1000 ms)

At this time, it resets the mass and mass-flow estimators' state.

It is possible that this "bleed" of pre-tare weights could have
triggered SAW with the older, 2-second tau estimation and led to
the logic that "ignores" the first few seconds for SAW. This logic
remains unchanged, with 5 and 2.5 second defaults (espresso, water)
found in earlier code.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>