Correct volume reporting, estimate 50/60 Hz, add on_shotvalue_available
Previously, the total volume was tracked as the summation of the flow rate
reported by the DE1, multiplied by the difference in time samples based on
a 16-bit counter in the DE1's report of half-cycles. Esiting code did not
unwrap the counter, resulting in negative values when it wrapped, roughly
every 10 minutes. It also used the hard-wired value of 50 Hz, which caused
overestimation for users on 60 Hz, impacting ::espresso_water_dispensed
and its dependencies.

At this time it does not appear that the line frequency is sent by the DE1
to the app. An estimator is provided that runs on connect of the DE1
and saves the result in $::settings(line_frequency). If that estimate
is present in ::settings() and is either 50 or 60, it is not run again.
If $::settings(line_frequency_always_estimate) is present and a true value,
it will be run on connect even if ::settings(line_frequency) seems valid.

The estimation process takes a bit over 10 seconds and will be delayed
if the Bluetooth queue is too long. This helps prevent misestimation
due to delays seen during the other code initializing the DE1 on connect.

Prior to the estimation prcess completing, a best-guess is made:
  50 Hz for 230 V (based on $::settings(heater_voltage))
  60 Hz for 230 V and kr as $::settings(language)
  60 Hz for 120 V
  50 Hz otherwise

Replacing this estimation process with a read from the DE1 firmware
would be a future, preferred approach, once available.

The raw, 16-bit counter is present in ShotSample(SampleTime),
consistent with the firmware definition, and replacing Timer.
As $::machine(hz) is hard-coded to 50, the numeric value is the same.
No uses of ShotSample(Timer) were found outside of the proc.

Within the proc, the wrap of the 16-bit counter is compensated for.
Existing logic around negative and too-large positive water increments
is retained, though is likely unnecessary. The negative increments
were likely due to a pair of samples crossing the wrap, resulting in
a "time" difference of (25 - 65536).

DE1 event on_shotvalue_available is added. No existing code
has been refactored to use this event.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>