misc: Update list of `files` to include those added in this branch
In proc make_de1_dir {srcdir destdirs} there is a list of `files` that is used
to copy files from a source to a destination.

Note: This proc appears to be unused at run time and potentially could be moved
to the de1app/misc/ directory, where it is used in `misc/makede1.tcl`

Updated this list to include the files added by this series of commits.

The list of `not_these` appears to be unused, but is retained in this commit.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>