de1: Rename to 'after_flow_complete_delay 5'
Use a more understandable name for the parameter that sets
the delay between the end of flow and the after_flow_complete event.

The default value of 5 seems conservative, based on tester's experience
with medium and lighter beans and profiles with moderate and long
soak times.

SIGNIFICANT END-USER CHANGES
============================

`after_flow_complete_delay` in settings now controls the delay
between the shot "ending" and the stop of the weight recording,
associated with the "Enjoy1" toast.

Reducing this delay is possible, with the potential impact of the
"final drink weight" not being quite right, as drops may still fall.

This parameter does not, at this time, impact the extraction itself.

`seconds_after_espresso_stop_to_continue_weighing` is now unused

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>